requirements.txt
rknn-toolkit
opencv-python
numpy
pyserial
rknn-api